requirements.txt
alembic==1.11.2
aniso8601==9.0.1
anitopy==2.1.1
APScheduler==3.10.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.12.2
better-exceptions==0.3.3
bs4==0.0.1
cacheout==0.14.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cn2an==0.5.21
colorama==0.4.4
colored==1.3.93
cryptography==41.0.3
cssselect==1.2.0
DBUtils==3.0.3
dateparser==1.1.8
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
Flask-Login==0.6.2
fast-bencode==1.1.3
flask-compress==1.14
flask-restx==1.1.0
flask-sock==0.6.0
func_timeout==4.3.5
greenlet==2.0.2
guessit==3.7.1
h11==0.14.0
humanize==4.7.0
icalendar==5.0.7
idna==3.4
influxdb==5.3.1
itsdangerous==2.1.2
iso639==0.1.4
jedi==0.19.0
Jinja2==3.1.2
jsonpath==0.82
jsonschema==4.19.0
loguru==0.7.0
lxml==4.9.3
libtorrent==2.0.9
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.5
openai==0.28.0
outcome==1.2.0
parse==1.19.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pikpakapi==0.1.3
pillow==10.0.0
PlexAPI==4.15.0
proces==0.1.6
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.4.1
PyMySQL==1.1.0
pyperclip==1.8.2
pypushdeer==0.0.3
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python_hosts==1.0.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
qbittorrent-api==2023.9.53
redis==3.5.3
redis-py-cluster==2.1.3
regex==2023.8.8
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
selenium==4.11.2
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
srt==3.5.3
terminal-layout==2.1.6
tqdm==4.66.1
traitlets==5.9.0
transmission-rpc==6.0.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver==3.5.3
urllib3==2.0.4
w3lib==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==4.0.0
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
zhconv==1.4.3
MechanicalSoup~=1.3.0
cloudscraper~=1.2.71
setuptools~=69.1.1